ecommerce_pipeline/requirements.txt
boto3==1.26.94
botocore==1.29.94
Faker==17.6.0
jmespath==1.0.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15